models.py for user and models2 to store allt classa to organise later in app
